[Pipelining] Fix TMA descriptor pipelining when only TMA store is pipelined
When we only pipeling TMA store we need to make sure we double buffer the TMA
descriptor.
We also need to make sure the store wait op is waiting until the store is done
and not only until the smem is read.
